+++ tests/testthat/test-filter.r
@@ -275,3 +275,10 @@ test_that("filter, slice and arrange preserves attributes (#1064)", {
   expect_equal( res, "this is important")
 
 })
+
+test_that("filter works with rowwise data (#1099)", {
+  df <- data_frame(First = c("string1", "string2"), Second = c("Sentence with string1", "something"))
+  res <- df %>% rowwise() %>% filter(grepl(First, Second, fixed = TRUE))
+  expect_equal( nrow(res), 1L)
+  expect_equal( df[1,], res)
+})